remove unnecessary things
