+++ package.json
@@ -30,8 +30,7 @@
     "web3-eth-accounts": "^4.1.2",
     "web3-eth-contract": "^4.5.0",
     "web3-types": "^1.6.0",
-    "web3-utils": "^4.3.0",
-    "web3-validator": "^2.0.6"
+    "web3-utils": "^4.3.0"
   },
   "devDependencies": {
     "@chainsafe/eslint-config": "^2.1.1",
